(nvim) Switch theme to a heavily customized version of `nightfox`
This commit deprecates your old theme `ryanlg/lights-off` that you have
used for years, and adds a similar monochrome theme that's based on
`EdenEast/nightfox.nvim`.

You've customized the theme to a T to make the majority of it
monochrome, while still preserving colors for certain aspects of the UI
(such as NvimTree, quick fixes, spell checks, etc).